requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
idna==3.6
kombu==5.3.7
Pillow==10.1.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.6
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
